# Remove shared_ptr inside VectorT