--- a/include/Basic/VectorT.hpp
+++ b/include/Basic/VectorT.hpp
@@ -16,7 +16,6 @@
 
 #include <vector>
 #include <sstream>
-#include <memory>
 #include <algorithm>
 #include <cmath>
 
@@ -41,38 +40,39 @@ public:
   typedef typename Vector::const_reverse_iterator   const_reverse_iterator;
 
 public:
-  inline VectorT()                                                    : _v(std::make_shared<Vector>()) { }
-  inline VectorT(const Vector& vec)                                   : _v(std::make_shared<Vector>(vec)) { }
-  inline VectorT(size_type count, const T& value = T())               : _v(std::make_shared<Vector>(count, value)) { }
+  inline VectorT()                                                    : _v() { }
+  inline VectorT(const Vector& vec)                                   : _v(vec) { }
+  inline VectorT(size_type count, const T& value = T())               : _v(count, value) { }
   template< class InputIt >
-  inline VectorT(InputIt first, InputIt last)                         : _v(std::make_shared<Vector>()) { _v->assign(first, last); }
+  inline VectorT(InputIt first, InputIt last)                         : _v() { _v.assign(first, last); }
   inline VectorT(const VectorT& other) = default;
 #ifndef SWIG
-  inline VectorT(std::initializer_list<T> init)                       : _v(std::make_shared<Vector>(init)) { }
+  inline VectorT(std::initializer_list<T> init)                       : _v(init) { }
   inline VectorT(VectorT&& other)                                      noexcept { _v.swap(other._v); }
 #endif
   inline ~VectorT() = default;
 
 #ifndef SWIG
-  inline operator const Vector&() const                               { return *_v; }
+  inline operator const Vector&() const                               { return _v; }
 #endif
 
-  inline Vector& getVector() const                                    { return *_v; }
-  inline Vector* getVectorPtr() const                                 { return _v.get(); }
+  inline Vector& getVector()                                    { return _v; }
+  inline const Vector& getVector() const                                    { return _v; }
+  inline const Vector* getVectorPtr() const                                 { return &_v; }
 
 #ifndef SWIG
-  inline VectorT& operator=(const Vector& vec)                        { _detach(); *_v = vec; return (*this); }
+  inline VectorT& operator=(const Vector& vec)                        { _detach(); _v = vec; return (*this); }
   inline VectorT& operator=(const VectorT& other)                     { _detach(); _v = other._v; return (*this); }
   inline VectorT& operator=(VectorT&& other)                           noexcept { _v.swap(other._v); return (*this); }
-  inline VectorT& operator=(std::initializer_list<T> init)            { _detach(); (*_v) = init; return (*this); }
+  inline VectorT& operator=(std::initializer_list<T> init)            { _detach(); _v = init; return (*this); }
 #endif
 
-  inline bool operator==(const VectorT& other) const                  { return *_v == *other._v; }
-  inline bool operator!=(const VectorT& other) const                  { return *_v != *other._v; }
-  inline bool operator <(const VectorT& other) const                  { return *_v  < *other._v; }
-  inline bool operator<=(const VectorT& other) const                  { return *_v <= *other._v; }
-  inline bool operator >(const VectorT& other) const                  { return *_v  > *other._v; }
-  inline bool operator>=(const VectorT& other) const                  { return *_v >= *other._v; }
+  inline bool operator==(const VectorT& other) const                  { return _v == other._v; }
+  inline bool operator!=(const VectorT& other) const                  { return _v != other._v; }
+  inline bool operator <(const VectorT& other) const                  { return _v  < other._v; }
+  inline bool operator<=(const VectorT& other) const                  { return _v <= other._v; }
+  inline bool operator >(const VectorT& other) const                  { return _v  > other._v; }
+  inline bool operator>=(const VectorT& other) const                  { return _v >= other._v; }
 
   // For SWIG users (size_type is not much appreciated)
   inline const T& getAt(int pos) const;
@@ -86,53 +86,53 @@ public:
   inline T& operator[](size_type pos);
 #endif
 
-  inline T& front()                                                   { _detach(); return _v->front(); }
-  inline const T& front() const                                       { return _v->front(); }
-  inline T& back()                                                    { _detach(); return _v->back(); }
-  inline const T& back() const                                        { return _v->back(); }
-  inline T* data()                                                    { _detach(); return _v->data(); }
-  inline const T* data() const                                        { return _v->data(); }
-  inline const T* constData() const                                   { return _v->data(); }
+  inline T& front()                                                   { _detach(); return _v.front(); }
+  inline const T& front() const                                       { return _v.front(); }
+  inline T& back()                                                    { _detach(); return _v.back(); }
+  inline const T& back() const                                        { return _v.back(); }
+  inline T* data()                                                    { _detach(); return _v.data(); }
+  inline const T* data() const                                        { return _v.data(); }
+  inline const T* constData() const                                   { return _v.data(); }
 
-  inline T* subdata(size_type i = 0)                                  { _detach(); return _v->data() + i; }
-  inline const T* subdata(size_type i = 0) const                      { return _v->data() + i; }
+  inline T* subdata(size_type i = 0)                                  { _detach(); return _v.data() + i; }
+  inline const T* subdata(size_type i = 0) const                      { return _v.data() + i; }
 
-  inline bool empty() const                                           { return _v->empty(); }
-  inline size_type size() const                                       { return _v->size(); }
-  inline void reserve(size_type new_cap)                              { _v->reserve(new_cap); }
-  inline size_type capacity() const                                   { return _v->capacity(); }
-  inline void clear()                                                 { _detach(); _v->clear(); }
+  inline bool empty() const                                           { return _v.empty(); }
+  inline size_type size() const                                       { return _v.size(); }
+  inline void reserve(size_type new_cap)                              { _v.reserve(new_cap); }
+  inline size_type capacity() const                                   { return _v.capacity(); }
+  inline void clear()                                                 { _detach(); _v.clear(); }
 
-  inline void insert(size_type i, const T& value)                     { _detach(); _v->insert(begin() + i, value); }
-  inline void insert(size_type i, size_type count, const T& value)    { _detach(); _v->insert(begin() + i, count, value); }
+  inline void insert(size_type i, const T& value)                     { _detach(); _v.insert(begin() + i, value); }
+  inline void insert(size_type i, size_type count, const T& value)    { _detach(); _v.insert(begin() + i, count, value); }
   inline iterator insert(const_iterator pos,
                          const_iterator first,
-                         const_iterator last )                        { _detach(); return _v->insert(pos, first, last); }
-  inline void remove(size_type i)                                     { _detach(); _v->erase(begin() + i); }
-  inline void remove(size_type i, size_type count)                    { _detach(); _v->erase(begin() + i, begin() + i + count); }
-  inline iterator erase( const_iterator pos )                         { _detach(); return _v->erase(pos); }
-  inline iterator erase( const_iterator first, const_iterator last)   { _detach(); return _v->erase(first, last); }
+                         const_iterator last )                        { _detach(); return _v.insert(pos, first, last); }
+  inline void remove(size_type i)                                     { _detach(); _v.erase(begin() + i); }
+  inline void remove(size_type i, size_type count)                    { _detach(); _v.erase(begin() + i, begin() + i + count); }
+  inline iterator erase( const_iterator pos )                         { _detach(); return _v.erase(pos); }
+  inline iterator erase( const_iterator first, const_iterator last)   { _detach(); return _v.erase(first, last); }
 
-  inline void push_back(const T& value)                               { _detach(); _v->push_back(value); }
-  inline void push_front(const T& value)                              { _detach(); _v->insert(begin(), value); }
+  inline void push_back(const T& value)                               { _detach(); _v.push_back(value); }
+  inline void push_front(const T& value)                              { _detach(); _v.insert(begin(), value); }
 #ifndef SWIG
-  inline void push_back(const T&& value)                              { _detach(); _v->push_back(value); }
-  inline void push_front(const T&& value)                             { _detach(); _v->insert(begin(), value); }
+  inline void push_back(const T&& value)                              { _detach(); _v.push_back(value); }
+  inline void push_front(const T&& value)                             { _detach(); _v.insert(begin(), value); }
 #endif
-  inline void resize(size_type count)                                 { if (count == size()) return; _detach(); _v->resize(count); }
-  inline void resize(size_type count, const T& value)                 { if (count == size()) return; _detach(); _v->resize(count, value); }
+  inline void resize(size_type count)                                 { if (count == size()) return; _detach(); _v.resize(count); }
+  inline void resize(size_type count, const T& value)                 { if (count == size()) return; _detach(); _v.resize(count, value); }
 
-  inline iterator begin()                                             { _detach(); return _v->begin(); }
-  inline const_iterator begin() const                                 { return _v->begin(); }
-  inline const_iterator cbegin() const                                { return _v->cbegin(); }
-  inline iterator end()                                               { _detach(); return _v->end(); }
-  inline const_iterator end() const                                   { return _v->end(); }
-  inline const_iterator cend() const                                  { return _v->cend(); }
+  inline iterator begin()                                             { _detach(); return _v.begin(); }
+  inline const_iterator begin() const                                 { return _v.begin(); }
+  inline const_iterator cbegin() const                                { return _v.cbegin(); }
+  inline iterator end()                                               { _detach(); return _v.end(); }
+  inline const_iterator end() const                                   { return _v.end(); }
+  inline const_iterator cend() const                                  { return _v.cend(); }
 
-  inline reverse_iterator rbegin()                                    { _detach(); return _v->rbegin(); }
-  inline const_reverse_iterator crbegin() const                       { return _v->crbegin(); }
-  inline reverse_iterator rend()                                      { _detach(); return _v->rend(); }
-  inline const_reverse_iterator crend() const                         { return _v->crend(); }
+  inline reverse_iterator rbegin()                                    { _detach(); return _v.rbegin(); }
+  inline const_reverse_iterator crbegin() const                       { return _v.crbegin(); }
+  inline reverse_iterator rend()                                      { _detach(); return _v.rend(); }
+  inline const_reverse_iterator crend() const                         { return _v.crend(); }
 
 #ifndef SWIG
   inline VectorT& operator<<(const T& value);
@@ -146,7 +146,7 @@ public:
   inline void assign(InputIt first, InputIt last)
   {
     _detach();
-    _v->assign(first, last);
+    _v.assign(first, last);
   }
 
   inline String toString(const AStringFormat* strfmt = nullptr) const;
@@ -159,7 +159,7 @@ public:
   inline void toTL() const {};
 
 protected:
-  std::shared_ptr<Vector> _v;
+  Vector _v;
 
 private:
   inline void _detach();
@@ -184,7 +184,7 @@ void VectorT<T>::setAt(int pos, const T& v)
 template <typename T>
 int VectorT<T>::length() const
 {
-  return static_cast<int>(_v->size());
+  return static_cast<int>(_v.size());
 }
 
 template <typename T>
@@ -211,7 +211,7 @@ const T& VectorT<T>::operator[](size_type pos) const
   // Unprotect operator[] ... as in std::vector library
   //  if (pos >= size())
   //    my_throw("VectorT<T>::operator[]: index out of range");
-  return _v->operator[](pos);
+  return _v.operator[](pos);
 }
 
 template<typename T>
@@ -221,7 +221,7 @@ T& VectorT<T>::operator[](size_type pos)
   //  if (pos >= size())
   //    my_throw("VectorT<T>::operator[]: index out of range");
   _detach();
-  return _v->operator[](pos);
+  return _v.operator[](pos);
 }
 #endif
 
@@ -270,9 +270,6 @@ void VectorT<T>::display(const AStringFormat* strfmt) const
 template<typename T>
 void VectorT<T>::_detach()
 {
-  if (_v.use_count() == 1)
-    return;
-  _v = std::make_shared<Vector>(*_v);
 }
 
 #ifndef SWIG