--- a/include/LinearOp/ProjComposition.hpp
+++ b/include/LinearOp/ProjComposition.hpp
@@ -13,6 +13,8 @@
 #include "gstlearn_export.hpp"
 #include "LinearOp/IProj.hpp"
 
+#include <memory>
+
 class GSTLEARN_EXPORT ProjComposition : public IProj
 {
 public: